requirements.txt
Pillow
win32-setctime
